# speed up H-matrix test